// generated/src/app/api-models/auth-claim-rule-select.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AuthClaimRuleSelect {
  lineOfBusiness: string;
  claimType: string;
  matchOrder: number;
  seqRuleId: number;
  columnName: string;
  columnOccurrence: number;
  columnType: string;
  operator: string;
  fromValue: string;
  thruValue: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}